requirements.txt
South==0.7.3
django-modeltranslation==0.3.2
sorl-thumbnail==11.05.2
django>=1.4
python-social-auth
requests
celery==3.1